# dify/tools/cs_openapi_tool.yaml
openapi: 3.0.0
info:
  title: 粤教服务智能客服 Agent - 业务闭环工具集
  description: 面向留学与教育服务机构的 OpenAPI 3.0 工具集，为 Dify Agent 提供课程推荐、活动查询、活动报名闭环、知识库检索与 FAQ 匹配能力。
  version: 1.0.0
servers:
  - url: http://host.docker.internal:8000/api/v1/cs
    description: Docker 容器访问 Windows 宿主机后端（Dify Docker 部署推荐）
  - url: http://127.0.0.1:8000/api/v1/cs
    description: 本地宿主机调试服务

paths:
  /courses/recommend:
    post:
      summary: 课程与项目个性化推荐
      operationId: recommendCourses
      description: 依据访客的学历层次（初中/中专/高中/大专/本科）、意向留学国家（德国/新加坡等）、预算上限等条件，智能推荐匹配度最高的前 1~3 个课程与升学项目。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                education_level:
                  type: string
                  description: 访客当前学历层次，例如：初中、中专、高中、大专、本科
                  example: 高中
                target_country:
                  type: string
                  description: 意向留学或培训国家，例如：德国、新加坡
                  example: 新加坡
                budget_max:
                  type: number
                  description: 留学总预算上限（人民币元）
                  example: 250000
                interest_keyword:
                  type: string
                  description: 感兴趣的方向或关键词，例如：免学费、双元制、带薪实习、专升本
                  example: 带薪实习
                recommend_limit:
                  type: integer
                  default: 3
                  description: 最多返回的推荐项目数量
      responses:
        '200':
          description: 推荐成功返回
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      is_matched:
                        type: boolean
                        example: true
                      match_count:
                        type: integer
                        example: 2
                      recommended_courses:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                            project_name:
                              type: string
                            category:
                              type: string
                            price:
                              type: number
                            duration:
                              type: string
                            description:
                              type: string
                            target_audience:
                              type: string
                            tags:
                              type: array
                              items:
                                type: string
                      recommendation_rationale:
                        type: string
                      follow_up_suggestion:
                        type: string

  /events:
    get:
      summary: 查询近期留学讲座与分享会活动
      operationId: listEvents
      description: 获取机构近期举办的线上或线下留学讲座、中德双元制宣讲会、新加坡专场等活动列表，包含开场时间、地点及剩余席位。
      responses:
        '200':
          description: 成功获取活动列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        event_name:
                          type: string
                        event_type:
                          type: string
                        start_time:
                          type: string
                        location:
                          type: string
                        max_participants:
                          type: integer
                        current_participants:
                          type: integer
                        has_available_seats:
                          type: boolean

  /events/register:
    post:
      summary: 讲座活动对话内一键报名预约
      operationId: registerEvent
      description: 登记访客姓名与手机/微信，为指定活动完成席位预定。后端自动进行名额剩余检查及手机号防重复报名拦截，报名成功后沉淀至数据库。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - event_id
                - customer_name
                - contact_info
              properties:
                event_id:
                  type: integer
                  description: 拟报名的活动 ID（通过 listEvents 获取）
                  example: 1
                customer_name:
                  type: string
                  description: 报名人姓名
                  example: 张三
                contact_info:
                  type: string
                  description: 手机号码或微信号（用于防重与通知发送）
                  example: "13800138000"
                remark:
                  type: string
                  description: 意向备注信息
                  example: 想了解德国双元制计算机方向
      responses:
        '200':
          description: 报名操作响应
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      is_success:
                        type: boolean
                        example: true
                      registration_id:
                        type: integer
                        example: 18
                      event_name:
                        type: string
                        example: 2026中德双元制带薪实训与欧标B1签约会
                      message:
                        type: string
                        example: 恭喜！您已成功预约活动席位

  /events/registrations:
    get:
      summary: 查询访客已报名的活动预约记录
      operationId: queryRegistrations
      description: 依据手机号码或联系方式，查询用户在系统后台已成功预约的所有讲座或活动记录，包含活动名称、开场时间及地点。
      parameters:
        - name: contact_info
          in: query
          required: true
          description: 预约时填写的手机号或联系方式
          schema:
            type: string
            example: "13800138000"
      responses:
        '200':
          description: 查询成功返回列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        registration_id:
                          type: integer
                        event_name:
                          type: string
                        start_time:
                          type: string
                        location:
                          type: string
                        customer_name:
                          type: string
                        contact_info:
                          type: string
                        status:
                          type: string

  /faq/match:
    post:
      summary: 36条标准FAQ高频问答秒级匹配
      operationId: matchFaq
      description: 对用户提问进行毫秒级高频标准 FAQ 检索，准确回答对公缴费银行账户、公司简称、学制、退费政策等标准问答。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - question
              properties:
                question:
                  type: string
                  description: 用户的咨询问题
                  example: 请问公司缴费的对公银行账户是多少？
                confidence_threshold:
                  type: number
                  default: 0.55
                  description: 置信度阈值 (0.0 - 1.0)
      responses:
        '200':
          description: FAQ 匹配结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: object
                    properties:
                      is_matched:
                        type: boolean
                        example: true
                      standard_question:
                        type: string
                      answer:
                        type: string
                      confidence_score:
                        type: number

  /kb/search:
    post:
      summary: 知识库混合检索
      operationId: searchKnowledgeBase
      description: 从企业信息、中德双元制、新加坡升学与留学政策切片库中混合检索最相关的知识片段与官方出处。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  description: 检索关键词或问题
                  example: 德国双元制实训津贴政策
                category:
                  type: string
                  description: 知识分类（company_info, business, policy, overseas_life）
                  example: business
                top_k:
                  type: integer
                  default: 3
      responses:
        '200':
          description: 知识库检索片段
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    example: 200
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        chunk_id:
                          type: integer
                        title:
                          type: string
                        content:
                          type: string
                        source_file:
                          type: string
                        relevance_score:
                          type: number